backend/requirements.txt
# Study Edge - Backend Requirements (Python 3.13 Compatible)
# This single file contains all necessary packages with Python 3.13 compatible versions

# Essential build system dependencies (required for Python 3.13)
setuptools>=68.0.0
wheel>=0.41.0
build>=1.0.0

# Core web framework and extensions
Flask>=3.0.0
Flask-JWT-Extended>=4.6.0
Flask-CORS>=4.0.1

# Database
Flask-PyMongo>=2.3.0
pymongo[srv]>=4.8.0

# Security and utilities
bcrypt>=4.1.0
python-dotenv>=1.0.0

# AWS services
boto3>=1.34.0
botocore>=1.34.0

# AI services
openai>=1.3.0

# File processing
openpyxl>=3.1.2

# Testing
pytest>=7.4.0

# Deployment and server
gunicorn>=22.0.0
Werkzeug>=3.0.0

# HTTP and utilities
python-multipart>=0.0.6
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3

# Text-to-speech and audio generation (required for LISTENING module)
gTTS>=2.5.0
pydub>=0.25.0

# Audio processing dependencies (Python 3.13 compatible versions)
# These versions have pre-built wheels for Python 3.13
numpy>=1.26.0
scipy>=1.11.0
librosa>=0.10.0

# Alternative audio processing packages (more Python 3.13 compatible)
soundfile>=0.12.0
resampy>=0.4.0

# Authentication and security
flask-bcrypt>=1.0.1

# Email service
brevo-python>=1.1.0

# Template engine
jinja2>=3.1.0

# Speech recognition and audio processing
SpeechRecognition>=3.10.0
ffmpeg-python>=0.2.0

# API and real-time communication
fastapi>=0.104.0
uvicorn>=0.24.0
python-socketio>=5.10.0
flask-socketio>=5.3.0
eventlet>=0.33.0

# Performance monitoring and optimization
psutil>=5.9.0
gunicorn[eventlet]>=22.0.0
waitress>=2.1.0

# Unix/Linux optimized workers for production
gunicorn[gevent]>=22.0.0
gevent>=23.0.0
uvicorn[standard]>=0.24.0

# Task scheduling
apscheduler>=3.10.0

# Additional Python 3.13 compatible packages
typing-extensions>=4.8.0
packaging>=23.0

# Note: For audio generation to work properly, you may also need to install system dependencies:
# - ffmpeg (for audio file processing)
# - portaudio (for audio recording)
# - libasound2-dev (for ALSA audio support on Linux)
# 
# On Ubuntu/Debian: sudo apt-get install ffmpeg portaudio19-dev libasound2-dev
# On CentOS/RHEL: sudo yum install ffmpeg portaudio-devel alsa-lib-devel
# On macOS: brew install ffmpeg portaudio
# On Windows: Download ffmpeg from https://ffmpeg.org/download.html
#
# Python 3.13 Compatibility:
# - This file uses versions that have pre-built wheels for Python 3.13
# - If you encounter build issues, consider using Python 3.11 or 3.12
# - All audio generation features for LISTENING module are included 